requirements.txt
databricks-sdk>=0.8.0
requests>=2.28.0
pyspark>=3.4.0
pandas>=1.5.0
python-dotenv>=0.19.0